backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==0.3.9
langchain-google-genai==2.0.8
langchain-chroma==0.2.0
langchain-community==0.3.9
chromadb==0.5.23
pymupdf==1.25.3
python-docx==1.1.2
python-multipart==0.0.9
pydantic-settings==2.6.1
sse-starlette==2.2.1
httpx==0.27.0
